refactored frame-specific logic into a different, separate component; moved utils into index file
