+++ teachertool/src/App.tsx
@@ -8,9 +8,8 @@ import Notifications from "./components/Notifications";
 import * as NotificationService from "./services/notificationService";
 import { postNotification } from "./transforms/postNotification";
 import { makeNotification } from "./utils";
-import { MakeCodeFrame } from "./components/makecodeFrame";
-import { isLocal, getEditorUrl } from "./utils/browserUtils";
 import DebugInput from "./components/DebugInput";
+import { EditorContainer } from "./components/EditorContainer";
 
 
 function App() {
@@ -32,34 +31,12 @@ function App() {
         }
     }, [ready]);
 
-    const onIframeLoaded = () => {
-        console.log("iframe loaded");
-    }
-
-    const onIframeClosed = () => {
-        console.log("iframe closed");
-    }
-
-    const createIFrameUrl = (): string => {
-        const editorUrl: string = isLocal() ? "http://localhost:3232/index.html#editor" : getEditorUrl((window as any).pxtTargetBundle.appTheme.embedUrl);
-
-        let url = editorUrl
-        if (editorUrl.charAt(editorUrl.length - 1) === "/" && !isLocal()) {
-            url = editorUrl.substr(0, editorUrl.length - 1);
-        }
-        url += `?controller=1&ws=browser&nocookiebanner=1`;
-        return url;
-    }
-
     return (
         <>
             <HeaderBar />
             <div className="app-container">
                 <DebugInput />
-                <MakeCodeFrame pageSourceUrl={createIFrameUrl()}
-                    onFrameOpen={onIframeLoaded}
-                    onFrameClose={onIframeClosed}
-                />
+                <EditorContainer />
             </div>
 
             <Notifications />
