--- teachertool/src/utils/browserUtils.ts
@@ -1,12 +1,0 @@
-export const isLocal = () => {
-    return window.location.hostname === "localhost";
-}
-
-export const getEditorUrl = (embedUrl: string) => {
-    if (!pxt.webConfig && (window as any).pxtConfig) pxt.setupWebConfig((window as any).pxtConfig);
-    if (pxt.webConfig?.targetUrl && pxt.webConfig?.relprefix) {
-        return pxt.webConfig.targetUrl + pxt.webConfig.relprefix.substr(0, pxt.webConfig.relprefix.length - 3);
-    }
-    const path = /\/([\da-zA-Z\.]+)(?:--)?/i.exec(window.location.pathname);
-    return `${embedUrl.replace(/\/$/, "")}/${path?.[1] || ""}`;
-}
