+++ teachertool/src/utils/index.ts
@@ -12,3 +12,16 @@ export function makeNotification(
         expiration: Date.now() + duration,
     };
 }
+
+export const isLocal = () => {
+    return window.location.hostname === "localhost";
+}
+
+export const getEditorUrl = (embedUrl: string) => {
+    if (!pxt.webConfig && (window as any).pxtConfig) pxt.setupWebConfig((window as any).pxtConfig);
+    if (pxt.webConfig?.targetUrl && pxt.webConfig?.relprefix) {
+        return pxt.webConfig.targetUrl + pxt.webConfig.relprefix.substr(0, pxt.webConfig.relprefix.length - 3);
+    }
+    const path = /\/([\da-zA-Z\.]+)(?:--)?/i.exec(window.location.pathname);
+    return `${embedUrl.replace(/\/$/, "")}/${path?.[1] || ""}`;
+}
